24th

+++ src/pages/MyServices.jsx
@@ -2,6 +2,7 @@ import React, { useContext, useEffect, useState } from 'react'
 import { AuthContext } from '../provider/AuthProvider';
 import { Link, useNavigate } from 'react-router';
 import axios from 'axios';
+import Swal from 'sweetalert2';
 
 const MyServices = () => {
 
@@ -17,17 +18,40 @@ const MyServices = () => {
     
 
     const handleDelete = (id) => {
+      Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (result.isConfirmed) {
       axios.delete(`http://localhost:3000/delete/${id}`)
       .then(res => {
-        console.log(res.data)
-        alert("Deletion successful.")
-        const filterData = MyServices.filter(service => service._id != id)
-        setMyServices(filterData)
-        navigation('/MyServices')
+        // console.log(res.data)
+        if(res.data.acknowledged){
+          const filterData = MyServices.filter(service => service._id != id)
+          setMyServices(filterData)
+          navigation('/MyServices')
+          Swal.fire({
+          title: "Deleted!",
+          text: "Your file has been deleted.",
+          icon: "success"
+        });
+        }
+        
+        
       })
       .catch(err => {
         console.log(err)
       })
+
+        
+      }
+    });
+      
     }
     
   return (
